+++ src/test/java/nextstep/ladder/domain/player/PlayerTest.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.player;
 
 import nextstep.ladder.domain.player.Player;
 import org.junit.jupiter.api.DisplayName;
